Cleanup and live reload after sync & stats
