+++ assets/css/app.css
@@ -6,3 +6,16 @@
 
 /* @import "./liveview.scss"; */
 @import url("https://fonts.googleapis.com/css2?family=Roboto+Slab&display=swap");
+
+/* Style scrollbar */
+
+/* Hide scrollbar for IE, Edge and Firefox */
+html {
+  -ms-overflow-style: none;  /* IE and Edge */
+  scrollbar-width: none;  /* Firefox */
+}
+
+/* Hide scrollbar for Chrome, Safari and Opera */
+html::-webkit-scrollbar {
+  display: none;
+}
